requirements.txt
agno==1.8.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
fastapi==0.121.0
filelock==3.18.0
gitdb==4.0.12
gitpython==3.1.45
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.1
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==1.26.4
openai==2.7.1
packaging==25.0
pandas==2.3.2
platformdirs==4.3.8
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
tomli==2.2.1
tqdm==4.67.1
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.15
uvicorn==0.38.0
virtualenv==20.32.0
youtube-transcript-api==1.2.3
yt-dlp==2025.10.14
pypdf
google-generativeai
lancedb
tantivy
google-genai
tavily-python
chromadb
